update to lalrpop 0.17
I have no idea what I'm doing here, just stabbing in the dark.  It
looks to me like lalrpop is hopefully safe to update:

    https://github.com/lalrpop/lalrpop/blob/master/RELEASES.md

The main thing appears to be to ensure that the dozen-or-so uses of
ParseError::UnrecognizedToken are not expected to handle EOF as well.
If they are, then they probably need to handle
ParseError::UnrecognizedEOF independently.

It looks to me like the version numbers and hashes found in Cargo.lock
is maybe also supposed to be updated, but I do not know how to do that
safely.  It looks like a generated file, and I'm generally averse to
committing generated files to revision control.