For Emacs users, set outline-regexp for some file types.
Do this in ".dir-locals.el", which was previously ignored by Git.

That file should not have been ignored as it is intended for settings
that the maintainers want to share with contributors.  For settings
that are NOT intended for sharing one should use ".dir-locals-2.el".

We tell Git to ignore the latter file instead of the former.

`outline-regexp` is a good example of a setting that should be shared,
with others.  It specifies the syntax used for section headings used
for certain files within this project.  Setting that value by itself
does not change how Emacs behaves.

Turing on `outline-minor-mode` on the other hand changes how the user
interacts with those headings.  Not everyone is going to like that, so
the mode should be enabled in ".dir-locals-2.el" instead.

`rustic-mode` is a fork of `rust-mode`, so we theme that too.