sq: Fix manpages where line starts with "'sq".
Three sequoia manpages are broken, with some text not appearing in the
manual output.  In particular, lines where the roff source starts with
"'sq" are not rendered.  This is because a leading ' on a line is
treated as a macro delimiter.

Without these changes, we see the following kind of warnings produced:

    0 $ LC_ALL=en_US.UTF-8 MANROFFSEQ='' MANWIDTH=80 man --warnings \
    > -E UTF-8 -l -Tutf8 -Z sq-packet-dump.1 > /dev/null
    troff: <standard input>:12: warning: macro 'sq' not defined
    0 $

See also https://lintian.debian.org/tags/groff-message.html.