sq: Align example usage with synopsis.
The synopsis clearly places TARGET-KEY *after* the options.  But the
example places it haphazardly within the options, which is harder to
read and understand.

Even if the original example works, the goal of the example should be
to make it easy to see what is happening in a standard usage, not to
show off the flexibility of the parser.