sq: Port command line handling to clap version 3
Change sq command line handling to use clap version 3, and adapt to
all the breaking changes.

Clap version 3 is a major new version with a number of breaking
changes. It also adds functionality to allow a structopt style
declarative way of defining command line syntax. We want to use that,
but first we need to port the old "builder style" of defining the
command line syntax to clap version 3. The change to use the "derive
style" comes later.

We now output a cargo:warning with paths for shell completion files
generated.

A subplot scenario was fixed: it used --signer-cert instead of
--signer-key. With clap v2 this was for some reason OK, but with v3
it's an error.

The semantics of clap version 2's .multiple function were hard to
understand and it was replaced with .multiple_occurences. Care was taken
to preserve the original intention regarding an argument's number of
occurrences and number of values.

There are some changes to help output (in src/sq-usage.rs). These are
mostly from upstream changes and we don't want to fork clap to change
them. In summary: FLAGS and OPTIONS are merged into just OPTIONS; the
layout of subcommand lists are a little different (split into two
lines); there is no "[--]" before filename arguments anymore; default
and allowed values for options are on a separate line now; --version
isn't repeated for every subcommand anymore; help is listed for each
subcommand separately.

In addition, we will help clap upstream fix a problem where the
help output doesn't have a "..." to indicate that an option may be used
several times.

Further, upstream has changed --help text to be of the form "Print
help", when it earlier was "Prints help". We will change our own help
texts to follow suite in a future commit. We don't do it in this commit,
to avoid making an even larger diff.